# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Neon House LED - Backend API",
        "description": "API REST para gestión de propuestas, clientes, blogs, productos y más",
        "contact": {
            "name": "Neon House LED",
            "url": "https://neonhouseled.com",
            "email": "support@neonhouseled.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Desarrollo local"
        },
        {
            "url": "http://localhost:3000/api",
            "description": "Producción (ejemplo)"
        }
    ],
    "paths": {
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Obtener datos del usuario autenticado",
                "description": "Retorna los datos del usuario actualmente autenticado, incluyendo roles y permisos",
                "operationId": "getAuthenticatedUser",
                "responses": {
                    "200": {
                        "description": "Datos del usuario autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/change-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cambiar contraseña del usuario autenticado",
                "description": "Permite al usuario autenticado cambiar su contraseña. Requiere la contraseña actual para validar.",
                "operationId": "changePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña cambida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChangePasswordSuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validación fallida o contraseña actual incorrecta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "La contraseña actual es incorrecta"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "currentPassword": [
                                                    "Campo requerido"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de validación inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cliente/{id}/propuestas": {
            "get": {
                "tags": [
                    "Propuestas"
                ],
                "summary": "Obtener propuestas de un cliente",
                "description": "Retorna todas las propuestas de un cliente específico incluyendo sus imágenes y videos",
                "operationId": "59f76ce04611dbe3b39717d6784652c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Propuestas encontradas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropuestasListSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado o no tiene propuestas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropuestaNotFoundResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de validación inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropuestaValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropuestaServerErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/propuestas": {
            "get": {
                "tags": [
                    "Propuestas"
                ],
                "summary": "Obtener todas las propuestas",
                "description": "Retorna todas las propuestas del sistema incluyendo información del cliente, imágenes y videos",
                "operationId": "275141d78415a86d093bf0e7cd8fd5a6",
                "responses": {
                    "200": {
                        "description": "Propuestas encontradas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropuestasAllListSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron propuestas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropuestaNotFoundResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropuestaServerErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "AuthErrorResponse": {
                "properties": {
                    "error": {
                        "description": "Mensaje de error de autenticación",
                        "type": "string",
                        "example": "No autorizado"
                    }
                },
                "type": "object"
            },
            "ChangePasswordRequest": {
                "required": [
                    "currentPassword",
                    "newPassword"
                ],
                "properties": {
                    "currentPassword": {
                        "description": "Contraseña actual del usuario",
                        "type": "string",
                        "format": "password",
                        "example": "currentPass123"
                    },
                    "newPassword": {
                        "description": "Nueva contraseña (mínimo 8 caracteres)",
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "example": "newPassword123"
                    }
                },
                "type": "object"
            },
            "ChangePasswordSuccessResponse": {
                "properties": {
                    "message": {
                        "description": "Mensaje de confirmación",
                        "type": "string",
                        "example": "Contraseña cambiada exitosamente"
                    }
                },
                "type": "object"
            },
            "PropuestaListItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Propuesta Premium"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción de la propuesta"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-15T10:30:00"
                    },
                    "id_cliente": {
                        "type": "integer",
                        "example": 5
                    },
                    "images": {
                        "description": "URLs de imágenes almacenadas",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uri"
                        },
                        "example": [
                            "/storage/cliente/5/propuestas/1/imagenes/image1.jpg"
                        ]
                    },
                    "videos": {
                        "description": "URLs de videos almacenados",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uri"
                        },
                        "example": [
                            "/storage/cliente/5/propuestas/1/videos/video1.mp4"
                        ]
                    }
                },
                "type": "object"
            },
            "PropuestaListWithClientItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_cliente": {
                        "type": "integer",
                        "example": 5
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Propuesta Premium"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción detallada de la propuesta"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-15T10:30:00"
                    },
                    "cliente_nombre": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "cliente_apellido": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "cliente_email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@example.com"
                    },
                    "images": {
                        "description": "URLs de imágenes almacenadas",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uri"
                        },
                        "example": [
                            "/storage/cliente/5/propuestas/1/imagenes/image1.jpg"
                        ]
                    },
                    "videos": {
                        "description": "URLs de videos almacenados",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uri"
                        },
                        "example": [
                            "/storage/cliente/5/propuestas/1/videos/video1.mp4"
                        ]
                    }
                },
                "type": "object"
            },
            "PropuestaNotFoundResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 404
                    },
                    "message": {
                        "type": "string",
                        "example": "No se encontraron propuestas"
                    },
                    "data": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
            "PropuestaServerErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 500
                    },
                    "message": {
                        "type": "string",
                        "example": "Error en el servidor"
                    }
                },
                "type": "object"
            },
            "PropuestaValidationErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 422
                    },
                    "message": {
                        "type": "object",
                        "example": {
                            "id_cliente": [
                                "El id_cliente es requerido",
                                "El id_cliente debe ser numérico"
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "PropuestasAllListSuccessResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200
                    },
                    "data": {
                        "description": "Array de propuestas con información del cliente",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PropuestaListWithClientItem"
                        }
                    }
                },
                "type": "object"
            },
            "PropuestasListSuccessResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "description": "Array de propuestas",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PropuestaListItem"
                        }
                    }
                },
                "type": "object"
            },
            "UserResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "empleado": {
                        "properties": {
                            "id_empleado": {
                                "type": "integer"
                            },
                            "nombre": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "errors": {
                        "type": "object",
                        "example": {
                            "currentPassword": [
                                "Campo requerido"
                            ],
                            "newPassword": [
                                "Mínimo 8 caracteres"
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login para obtener el token Bearer jwt",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Propuestas",
            "description": "Propuestas"
        }
    ]
}